refactor livechat codes
