+++ packages/core/src/App/Components/Elements/LiveChat/index.js
@@ -1,0 +1,3 @@
+import LiveChat from './live-chat.jsx';
+
+export default LiveChat;
