+++ packages/core/src/App/Containers/Layout/footer.jsx
@@ -13,7 +13,7 @@ import {
     ToggleFullScreen,
     ToggleSettings,
 } from 'App/Components/Layout/Footer';
-import LiveChat from 'App/Components/Elements/live-chat.jsx';
+import LiveChat from 'App/Components/Elements/LiveChat';
 import { connect } from 'Stores/connect';
 import ServerTime from '../server-time.jsx';
